+++ packages/camera/camera_web/example/integration_test/camera_test.dart
@@ -53,7 +53,7 @@ void main() {
     });
 
     setUpAll(() {
-      registerFallbackValue<CameraOptions>(MockCameraOptions());
+      registerFallbackValue(MockCameraOptions());
     });
 
     group('initialize', () {
@@ -1477,7 +1477,7 @@ void main() {
     });
 
     group('dispose', () {
-      testWidgets('resets the video element\'s source',
+      testWidgets("resets the video element's source",
           (WidgetTester tester) async {
         final Camera camera = Camera(
           textureId: textureId,